Extend `--apple_bitcode` to take an optional platform and allow multiple occurrences.
This change changes `--apple_bitcode` in a backward-compatible way to accept `--apple_bitcode=[platform=]mode`, where `platform` is `ios`, `macos`, `tvos`, or `watchos`. If a platform is specified, the given Bitcode mode will be applied specifically to that platform; if omitted, the option behaves as before, applying the Bitcode mode to all platforms. The option can be listed multiple times, with later ones winning, so a default can be set for all platforms and then overridden for specific ones.

This new format can be useful, for example, when building watchOS apps that must be distributed with Bitcode but where you do not want to build the companion iOS app with it enabled. To achieve this, pass `--apple_bitcode=watchos=embedded` to your build invocation.

RELNOTES: `--apple_bitcode` now takes an optional platform and only applies the Bitcode mode to that platform if present. The option may be provided multiple times.
PiperOrigin-RevId: 324023106