[Bazel] Cherry-pick [Log] Surface error log file for manifest merging issue
Summary:
Code taken from:
https://github.com/bazelbuild/bazel/pull/11691/commits/9a4afa661b2eca9abb4db3d17f76b9bf21d47f2f

Had to be reconciled with:
https://code.uberinternal.com/diffusion/MOBAZXL/browse/4.2.0-uber/src/main/java/com/google/devtools/build/lib/worker/WorkerSpawnRunner.java$354

It looks like the only material difference is that in the PR, recordingStreamMessage is in the logText() whereas it is not in the message.
The logSizeLimit() was also removed.

Test Plan:
We accept correctness if we're able to build bazel and run our code
{F602470913}

Built uberlite:
{F602496037}

Reviewers: oliviern

Reviewed By: oliviern

Revert Plan: Git revert

JIRA Issues: JAVADEVX-6011

Differential Revision: https://code.uberinternal.com/D7233455