[Bazel][Draft] Bazel Modification For Controlling Memory
Summary:
This change allows us to control individual actions via their mneonics

https://docs.google.com/document/d/1bxmBsUg9PTCrAtRnazH3rBUBMwnM6zC-YTU7l-CpaxY/edit

In general, this will help us avoid the Kotlin compiling error. It will help in avoiding OOM for dexing.

Why do we not put the estimate directly in the action itself? Why do we need to override it in resource manager?

If you look up the usage of .setMneonic(), you will see that they all exist in packages and classes and not every class has access to the options.

It seems very complicated to give every place that spawnActionBuilder gets created and used access to the override options. The fix becomes simpler when we call it before resourceManager gets invoked.

You are also not guaranteed where/when resoucebuilder gets called in the future. This can cause additional breakage in the option if the developer forgets to write it there.

Misc.

I don’t know a good way to handle the numbersparsingexception other than failing execution if/when it happens. It definitely seems weird sticking an optionsparsingexception all the way deep in the execution phase instead of the analysis phase. This error should also be reported to the user which currently it doesn’t unless you use --verbose_failure.

Test Plan:
./bazelw build uberliteDebug
{F635386623}

For failure:
./bazelw build riderDebug

{F635384631}

+
+build:linux --mnemonic_resource_override=AndroidDexer=1400ddx4
+

Reviewers: #ldap_mobile-platform-android!

Revert Plan: Git revert

Differential Revision: https://code.uberinternal.com/D7715239