[Java] Merge spring and cxf/bus-extensions.txt files
During the creation of a fat jar, several files on the classpath can
have the same name. In some situations, these files should be merged
together. In this commit, we add Concatenators for every file under
META-INF/spring except spring.factories, so that these files are merged
together.

spring.factories is excluded because it needs special processing. This
is a properties file, so it needs to be properly merged - if several
spring.factories have the same key, and the files are simply
concatenated, only one instance of the key will win. Instead, the
values from each file need to be merged together under one key.

This commit also implementes a sublcass of Concatenator called the
MergingPropertyCombiner. This combiner will handle the file in the way
described above. We also instantiate a MergingPropertyCombiner for
spring.factories so that the files are properly merged, allowing the
spring application to begin.

Additionally, we add a Concatenator for META-INF/cxf/bus-extensions.txt,
which is needed for the same reasons as the spring.factories files.