[Java] Update lite binary runscript template to call exec
Some lite binaries experienced issues with graceful shutdown. This was
due to the runscript not calling exec on $JAVABIN if the classpath was
too large and Bazel had to create a manifest jar.

In this commit, we update this line to call exec, allowing for signals
to be propagated to the java process correctly.