[Coverage] Fix coverage for large targets
When a classpath is very large, Bazel will create a manifest jar with
the classpath as an entry in the manifest.

To get the sources for coverage, Bazel checks whether it needs to
reverse the process to find the classpath. However, part of this logic
checks whether the manfiest jar is the only jar on the classpath
(urls.length == 1). This is not true in our case - there are two jars on
the classpath, with the first being the manifest jar.

In this commit, we update the logic to check whether the first jar on
the classpath is a manfiest jar by looking for the suffix on the jar.
This allows Bazel to properly reverse the process when the classpath is
large, meaning sources are found, and coverage is properly found again.